Split card styles, fix query cache invalidation on tx
